+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Allow Taskfiles starting with lowercase characters (#947, #1221 by @pd93).
   - e.g. `taskfile.yml`, `taskfile.yaml`, `taskfile.dist.yml` &
     `taskfile.dist.yaml`
-- Bug fixed were made to the
+- Bug fixes were made to the
   [npm installation method](https://taskfile.dev/installation/#npm). (#1190, by
   @sounisi5011).
 - Added the [gentle force experiment](https://taskfile.dev/experiments) as a